Replace Signal class with broadcast/handle system
The AS3-style Signal objects are no longer used.
There is instead a class called Signal in Scene.ooc which is a unit of communication between entities. Signals have a generic 'data' field which can be retrieved through 'sig get(Type)'. This makes them more like messages, than the callback lists that they used to be.

To send a signal to other entities, you use the 'broadcast' method, which can be directed at every entity in the scene, at all entities of a certain type, at all entities in an array of types, or at all entities in a given List.

Signals created through entity broadcast() will run through the Scene's handle() method which you can override, allowing them to be intercepted, or for the game to take global action on them.